requirements.txt
Flask
gunicorn
selenium
chromedriver-binary
google-cloud-storage
google-cloud-bigquery
google-api-python-client
google-cloud-secret-manager
pandas
pandas_gbq
webdriver-manager
